requirements.txt
# пустая строка в начале
pytest
selenium
allure-pytest
webdriver-manager